Soluciones de sintaxis, documentacion del codigo

+++ app/Http/Controllers/IdeaImageController.php
@@ -1,21 +1,25 @@
 <?php
+
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Idea;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
-
 class IdeaImageController extends Controller
 {
     public function destroy(Idea $idea)
-
     {
         Gate::authorize('workWith', $idea);
-        Storage::disk('public')->delete($idea->image_path);
-        $idea->update(['image_path' => null]);
 
-        return back();
+        if ($idea->image_path) {
+            Storage::disk('public')->delete($idea->image_path);
+
+            $idea->update([
+                'image_path' => null,
+            ]);
+        }
+
+        return back()->with('success', 'Image removed!');
     }
 }
